requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
streamlit==1.27.2
langchain==0.0.183
openai==0.27.7
tiktoken==0.4.0
streamlit-chat==0.1.1
faiss-cpu==1.7.4